# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PA MedTrack API - Pennsylvania PCH/ALR Compliance Tracking
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /facilities:
    get:
      operationId: listFacilities
      tags: [facilities]
      summary: List facilities
      parameters:
        - name: organizationId
          in: query
          schema:
            type: integer
        - name: facilityType
          in: query
          schema:
            type: string
        - name: isActive
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of facilities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Facility"
    post:
      operationId: createFacility
      tags: [facilities]
      summary: Create a facility
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFacilityBody"
      responses:
        "201":
          description: Facility created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Facility"
  /facilities/{id}:
    get:
      operationId: getFacility
      tags: [facilities]
      summary: Get facility by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Facility
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Facility"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateFacility
      tags: [facilities]
      summary: Update facility
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFacilityBody"
      responses:
        "200":
          description: Updated facility
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Facility"
    delete:
      operationId: deleteFacility
      tags: [facilities]
      summary: Delete facility
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /facilities/{id}/compliance-summary:
    get:
      operationId: getFacilityComplianceSummary
      tags: [facilities]
      summary: Get compliance summary for a facility
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Facility compliance summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FacilityComplianceSummary"

  # Employees
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List employees
      parameters:
        - name: facilityId
          in: query
          schema:
            type: integer
        - name: organizationId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: administersMedications
          in: query
          schema:
            type: boolean
        - name: trainerStatus
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create an employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeBody"
      responses:
        "201":
          description: Employee created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get employee by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateEmployee
      tags: [employees]
      summary: Update employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEmployeeBody"
      responses:
        "200":
          description: Updated employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /employees/{id}/compliance-summary:
    get:
      operationId: getEmployeeComplianceSummary
      tags: [employees]
      summary: Get compliance summary for an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee compliance summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeComplianceSummary"

  # Training Types
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Facility:
      type: object
      properties:
        id:
          type: integer
        organizationId:
          type: integer
        name:
          type: string
        facilityType:
          type: string
          enum: [PCH, ALR]
        licenseNumber:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        zip:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        administratorName:
          type: ["string", "null"]
        administratorEmail:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - organizationId
        - name
        - facilityType
        - isActive
        - createdAt
        - updatedAt
    CreateFacilityBody:
      type: object
      properties:
        organizationId:
          type: integer
        name:
          type: string
        facilityType:
          type: string
          enum: [PCH, ALR]
        licenseNumber:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        zip:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        administratorName:
          type: ["string", "null"]
        administratorEmail:
          type: ["string", "null"]
      required:
        - organizationId
        - name
        - facilityType
    UpdateFacilityBody:
      type: object
      properties:
        name:
          type: string
        facilityType:
          type: string
          enum: [PCH, ALR]
        licenseNumber:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        zip:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        administratorName:
          type: ["string", "null"]
        administratorEmail:
          type: ["string", "null"]
        isActive:
          type: boolean
    FacilityComplianceSummary:
      type: object
      properties:
        facilityId:
          type: integer
        facilityName:
          type: string
        facilityType:
          type: string
        totalEmployees:
          type: integer
        medAdminStaff:
          type: integer
        compliantCount:
          type: integer
        dueSoonCount:
          type: integer
        expiredCount:
          type: integer
        missingCount:
          type: integer
        complianceScore:
          type: number
        practicumsDue:
          type: integer
        annualHoursIncomplete:
          type: integer
      required:
        - facilityId
        - facilityName
        - facilityType
        - totalEmployees
        - medAdminStaff
        - compliantCount
        - dueSoonCount
        - expiredCount
        - missingCount
        - complianceScore
        - practicumsDue
        - annualHoursIncomplete
    Employee:
      type: object
      properties:
        id:
          type: integer
        organizationId:
          type: integer
        facilityId:
          type: integer
        employeeNumber:
          type: ["string", "null"]
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        hireDate:
          type: ["string", "null"]
        terminationDate:
          type: ["string", "null"]
        jobTitle:
          type: string
        department:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, terminated, on_leave]
        administersMedications:
          type: boolean
        trainerStatus:
          type: boolean
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - organizationId
        - facilityId
        - firstName
        - lastName
        - jobTitle
        - status
        - administersMedications
        - trainerStatus
        - createdAt
        - updatedAt
    CreateEmployeeBody:
      type: object
      properties:
        organizationId:
          type: integer
        facilityId:
          type: integer
        employeeNumber:
          type: ["string", "null"]
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        hireDate:
          type: ["string", "null"]
        jobTitle:
          type: string
        department:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, terminated, on_leave]
        administersMedications:
          type: boolean
        trainerStatus:
          type: boolean
        notes:
          type: ["string", "null"]
      required:
        - organizationId
        - facilityId
        - firstName
        - lastName
        - jobTitle
        - status
        - administersMedications
        - trainerStatus
    UpdateEmployeeBody:
      type: object
      properties:
        employeeNumber:
          type: ["string", "null"]
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        hireDate:
          type: ["string", "null"]
        terminationDate:
          type: ["string", "null"]
        jobTitle:
          type: string
        department:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, terminated, on_leave]
        administersMedications:
          type: boolean
        trainerStatus:
          type: boolean
        notes:
          type: ["string", "null"]
        facilityId:
          type: integer
    EmployeeComplianceSummary:
      type: object
      properties:
        employeeId:
          type: integer
        employeeName:
          type: string
        status:
          type: string
        administersMedications:
          type: boolean
        trainerStatus:
          type: boolean
        trainingRecords:
          type: array
          items:
            $ref: "#/components/schemas/TrainingRecord"
        practicums:
          type: array
          items:
            $ref: "#/components/schemas/Practicum"
        annualHours:
          type: array
          items:
            $ref: "#/components/schemas/TrainingHourBucket"
        documents:
          type: array
          items:
            $ref: "#/components/schemas/TrainingDocument"
        overallStatus:
          type: string
      required:
        - employeeId
        - employeeName
        - status
        - administersMedications
        - trainerStatus
        - trainingRecords
        - practicums
        - annualHours
        - documents
        - overallStatus
    TrainingType:
      type: object
      properties:
        id:
          type: integer
        organizationId:
          type: ["integer", "null"]
        code:
          type: string
        name:
          type: string
        category:
          type: string
        description:
          type: ["string", "null"]
        appliesToFacilityType:
          type: string
          enum: [PCH, ALR, BOTH]
        appliesToAdministersMeds:
          type: ["boolean", "null"]
        appliesToTrainers:
          type: ["boolean", "null"]
        renewalIntervalDays:
          type: ["integer", "null"]
        warningDaysDefault:
          type: integer
        documentRequired:
          type: boolean
        isSystemDefault:
          type: boolean
        isActive:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - code
        - name
        - category
        - appliesToFacilityType
        - warningDaysDefault
        - documentRequired
        - isSystemDefault
        - isActive
        - sortOrder
        - createdAt
        - updatedAt
...
    TrainingRecord:
      type: object
      properties:
        id:
          type: integer
        organizationId:
          type: integer
        facilityId:
          type: integer
        employeeId:
          type: integer
        trainingTypeId:
          type: integer
        completionDate:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        status:
          type: string
          enum: [compliant, due_soon, expired, missing, not_applicable, pending_review]
        trainerName:
          type: ["string", "null"]
        trainerCredentials:
          type: ["string", "null"]
        trainingProvider:
          type: ["string", "null"]
        certificateNumber:
          type: ["string", "null"]
        score:
          type: ["number", "null"]
        hours:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        documentRequired:
          type: boolean
        completionMethod:
          type: ["string", "null"]
          enum: [in_person, online, hybrid, manual_entry, null]
        verifiedByUserId:
          type: ["integer", "null"]
        verifiedAt:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        trainingType:
          $ref: "#/components/schemas/TrainingType"
      required:
        - id
        - organizationId
        - facilityId
        - employeeId
        - trainingTypeId
        - status
        - documentRequired
        - createdAt
        - updatedAt
        - trainingType
...
    Practicum:
      type: object
      properties:
        id:
          type: integer
        organizationId:
          type: integer
        facilityId:
          type: integer
        employeeId:
          type: integer
        practicumYear:
          type: integer
        completionDate:
          type: ["string", "null"]
        observedBy:
          type: ["string", "null"]
        marReviewCompleted:
          type: boolean
        directObservationCompleted:
          type: boolean
        remediationRequired:
          type: boolean
        remediationNotes:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        status:
          type: string
          enum: [compliant, due_soon, expired, missing]
        verifiedByUserId:
          type: ["integer", "null"]
        verifiedAt:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - organizationId
        - facilityId
        - employeeId
        - practicumYear
        - marReviewCompleted
        - directObservationCompleted
        - remediationRequired
        - status
        - createdAt
        - updatedAt
...
    TrainingDocument:
      type: object
      properties:
        id:
          type: integer
        organizationId:
          type: integer
        facilityId:
          type: integer
        employeeId:
          type: ["integer", "null"]
        trainingRecordId:
          type: ["integer", "null"]
        fileName:
          type: string
        fileUrl:
          type: string
        fileType:
          type: string
        fileSize:
          type: ["integer", "null"]
        uploadedByUserId:
          type: ["integer", "null"]
        documentType:
          type: string
          enum: [certificate, roster, practicum_form, transcript, other]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - organizationId
        - facilityId
        - fileName
        - fileUrl
        - fileType
        - documentType
        - createdAt
...
    TrainingHourBucket:
      type: object
      properties:
        id:
          type: integer
        organizationId:
          type: integer
        facilityId:
          type: integer
        employeeId:
          type: integer
        trainingYear:
          type: integer
        requiredHours:
          type: number
        completedHours:
          type: number
        status:
          type: string
          enum: [compliant, due_soon, incomplete, expired]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - organizationId
        - facilityId
        - employeeId
        - trainingYear
        - requiredHours
        - completedHours
        - status
        - createdAt
        - updatedAt
...