# lib/api-zod/src/generated/types/practicum.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PA MedTrack API - Pennsylvania PCH/ALR Compliance Tracking
 * OpenAPI spec version: 0.1.0
 */
import type { PracticumStatus } from "./practicumStatus";

export interface Practicum {
  id: number;
  organizationId: number;
  facilityId: number;
  employeeId: number;
  practicumYear: number;
  /** @nullable */
  completionDate?: string | null;
  /** @nullable */
  observedBy?: string | null;
  marReviewCompleted: boolean;
  directObservationCompleted: boolean;
  remediationRequired: boolean;
  /** @nullable */
  remediationNotes?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  dueDate?: string | null;
  status: PracticumStatus;
  /** @nullable */
  verifiedByUserId?: number | null;
  /** @nullable */
  verifiedAt?: string | null;
  createdAt: Date;
  updatedAt: Date;
}